OneDrive/Documents/ProjetIA/AgentGEE/backend/requirements.txt
earthengine-api>=0.1.348
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
python-dotenv>=0.19.0
gunicorn>=20.1.0